changed folder from models to src need to refactor group.py
